Move instruction and loop tests to reflect new package structure
