+++ tests/hardware/setup_tests.py
@@ -9,7 +9,7 @@ from qctoolkit.hardware.setup import HardwareSetup, PlaybackChannel, MarkerChann
 from tests.pulses.sequencing_dummies import DummyWaveform
 
 from tests.hardware.dummy_devices import DummyAWG, DummyDAC
-from tests.hardware.program_tests import get_two_chan_test_block, WaveformGenerator
+from tests._program.loop_tests import get_two_chan_test_block, WaveformGenerator
 
 
 class SingleChannelTests(unittest.TestCase):
